# perf: use popcnt in tr_bitfield