Same as [PR #2938](https://github.com/transmission/transmission/issues/2938) but instead of relying on [libpopcnt](https://github.com/transmission/libpopcnt) just makes new header file [tr-popcnt.h](https://github.com/transmission/transmission/compare/main...goldsteinn:submittable-hdr?expand=1#diff-ef11ec5a5d3fd4a66efbaafcd1ff9933faa5db09d766ec5f5bac20556ed13131) that accomplishes the same logic (but better for smaller types such as the `uint8_t` currently used).

Some notes: Tested with GCC11 on ubuntu20.04

Tested `tr_popcnt` independently with GCC11/Clang12 on ubuntu20.04 but the MSVC  preprocessor logic is untested. 
All `count` implementations are tested, however.

All tests either pass or are skipped:

```
$> libtransmission-test
...
[----------] 5 tests from Bitfield
[ RUN      ] Bitfield.count
[       OK ] Bitfield.count (1156 ms)
[ RUN      ] Bitfield.ctorFromFlagArray
[       OK ] Bitfield.ctorFromFlagArray (0 ms)
[ RUN      ] Bitfield.setRaw
[       OK ] Bitfield.setRaw (0 ms)
[ RUN      ] Bitfield.bitfields
[       OK ] Bitfield.bitfields (0 ms)
[ RUN      ] Bitfield.hasAllNone
[       OK ] Bitfield.hasAllNone (0 ms)
[----------] 5 tests from Bitfield (1156 ms total)
...
[ RUN      ] CompletionTest.createPieceBitfield
[       OK ] CompletionTest.createPieceBitfield (2 ms)
...
[----------] Global test environment tear-down
[==========] 270 tests from 37 test suites ran. (26791 ms total)
[  PASSED  ] 252 tests.
[  SKIPPED ] 18 tests, listed below:
[  SKIPPED ] JSON/JSONTest.testElements/1
[  SKIPPED ] JSON/JSONTest.testElements/2
[  SKIPPED ] JSON/JSONTest.testElements/3
[  SKIPPED ] JSON/JSONTest.testUtf8/1
[  SKIPPED ] JSON/JSONTest.testUtf8/2
[  SKIPPED ] JSON/JSONTest.testUtf8/3
[  SKIPPED ] JSON/JSONTest.test1/1
[  SKIPPED ] JSON/JSONTest.test1/2
[  SKIPPED ] JSON/JSONTest.test1/3
[  SKIPPED ] JSON/JSONTest.test2/1
[  SKIPPED ] JSON/JSONTest.test2/2
[  SKIPPED ] JSON/JSONTest.test2/3
[  SKIPPED ] JSON/JSONTest.test3/1
[  SKIPPED ] JSON/JSONTest.test3/2
[  SKIPPED ] JSON/JSONTest.test3/3
[  SKIPPED ] JSON/JSONTest.unescape/1
[  SKIPPED ] JSON/JSONTest.unescape/2
[  SKIPPED ] JSON/JSONTest.unescape/3
```