@goldsteinn 2022-04-20
Just want to add. Comparing the `libpopcnt` version you where going to use vs. main:

```
bits, Time Main, Time Libpopcnt, Time Main / Time Libpopcnt
   1,     0.147,          0.183,                      0.803
   8,     0.145,          0.209,                      0.693
  15,     0.263,          0.210,                      1.252
  22,     0.256,          0.206,                      1.248
  29,     0.307,          0.196,                      1.564
  36,     0.361,          0.207,                      1.740
  43,     0.402,          0.234,                      1.715
  50,     0.458,          0.248,                      1.848
  57,     0.507,          0.257,                      1.976
  64,     0.506,          0.249,                      2.028
 320,     2.180,          0.285,                      7.653
 576,     4.053,          0.323,                     12.533
 832,     5.663,          0.357,                     15.847
1088,     7.316,          0.401,                     18.256
1344,     8.956,          0.448,                     19.994
1600,    10.620,          0.497,                     21.352
1856,    12.262,          0.551,                     22.260
2112,    13.927,          0.596,                     23.354
2368,    15.547,          0.667,                     23.322
2624,    17.174,          0.743,                     23.111
2880,    18.887,          0.839,                     22.523
3136,    20.608,          0.986,                     20.902
3392,    22.168,          1.014,                     21.858
3648,    23.836,          1.079,                     22.086
3904,    25.518,          1.128,                     22.627
4160,    27.329,          0.614,                     44.523
4416,    28.761,          0.648,                     44.372
4672,    30.508,          0.670,                     45.546
4928,    32.272,          0.693,                     46.585
5184,    33.956,          0.711,                     47.780
5440,    35.505,          0.727,                     48.835
5696,    37.093,          0.745,                     49.756
5952,    38.850,          0.763,                     50.933
6208,    40.472,          0.780,                     51.855
6464,    42.100,          0.798,                     52.727
6720,    43.661,          0.817,                     53.414
6976,    45.521,          0.837,                     54.391
7232,    47.375,          0.856,                     55.358
7488,    48.858,          0.874,                     55.930
7744,    50.523,          0.891,                     56.704
8000,    52.098,          0.911,                     57.166
8192,    53.388,          0.925,                     57.715
```

Not arguing to use `libpopcnt`. But if `bitfield` is expected to be large
there is significant benefit to implementing a more optimized loop.