@ckerr 2022-04-21
> What is the expected size of the bitfield btw? If we are expecting large bitfields could make the code significantly faster. I made more original patch so there would be no regressions but even a simple unrolling of the loop to better hide the latency gets significant gain (although costs a bit for small sizes)

There is no real upper bound. Yes, these bitfields can potentially be very large when they are representing mega-torrents, so if you want to widen / unroll / etc. and have benchmarks to demonstrate the improvements I'd be happy  to consider for inclusion.