@goldsteinn 2022-04-21
> The macOS CI failure is not a blocker for this PR but I'm going to leave it open for now since @goldsteinn has indicated there could be more commits

Will be a bit before I have time to circle back to optimize the loop. Would prefer this go out first and can