@ckerr 2022-04-21
OK just a quick checkin -- @goldsteinn I saw a comment earlier (now deleted?) that you'd rather this PR go in as-is and then have another followup to improve speed. If that's still the case, lmk and I'll merge this one.